pyproject.toml
[tool.poetry]
name = "my_board"
version = "0.1.0"
description = ""
authors = ["Kangur <urbanek.kamil@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Jinja2 = "^3.1.2"
pyaml = "^21.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
